requirements.txt
streamlit
google-generativeai
pydantic
docxtpl
python-docx
python-dotenv
watchdog
